Slack issues fixes. 1. for transaction response model.
2. Update parcel and weight_unit required message.

+++ terminal_androidsdk/src/main/java/com/terminal/terminal_androidsdk/core/model/Packaging.kt
@@ -10,11 +10,11 @@ class Packaging (
     val  name:String,
     val size_unit:String,
     val type:String,
-    val height:Int,
-    val length:Int,
-    val weight:Int,
+    val height:Double,
+    val length:Double,
+    val weight:Double,
     val weight_unit :String,
-    val width:Int,
+    val width:Double,
     val description:String = ""
 )
