+++ terminal_androidsdk/src/main/java/com/terminal/terminal_androidsdk/core/model/component_shipment/Content.kt
@@ -9,7 +9,7 @@ data class Content (
   @SerializedName("description"           ) @Expose var description           : String?             = null,
   @SerializedName("packages"              ) @Expose  var packages              : ArrayList<Packages> = arrayListOf(),
   @SerializedName("declaredValueCurrency" )  @Expose var declaredValueCurrency : String?             = null,
-  @SerializedName("declaredValue"         )  @Expose var declaredValue         : Int?                = null,
+  @SerializedName("declaredValue"         )  @Expose var declaredValue         : Double?                = null,
   @SerializedName("incoterm"              ) @Expose  var incoterm              : String?             = null,
   @SerializedName("isCustomsDeclarable"   ) @Expose  var isCustomsDeclarable   : Boolean?            = null,
   @SerializedName("unitOfMeasurement"     ) @Expose  var unitOfMeasurement     : String?             = null,
