+++ terminal_androidsdk/src/main/java/com/terminal/terminal_androidsdk/core/model/component_shipment/DefaultParcel.kt
@@ -7,6 +7,6 @@ import com.google.gson.annotations.SerializedName
 data class DefaultParcel (
 
   @SerializedName("packaging_dimension" ) @Expose var packagingDimension : PackagingDimension? = PackagingDimension(),
-  @SerializedName("parcel_total_weight" )  @Expose var parcelTotalWeight  : Int?                = null
+  @SerializedName("parcel_total_weight" )  @Expose var parcelTotalWeight  : Double?                = null
 
 )
