+++ terminal_androidsdk/src/main/java/com/terminal/terminal_androidsdk/core/model/component_shipment/Items.kt
@@ -9,8 +9,8 @@ data class Items (
   @SerializedName("description" ) @Expose var description : String? = null,
   @SerializedName("name"        )  @Expose var name        : String? = null,
   @SerializedName("currency"    )  @Expose var currency    : String? = null,
-  @SerializedName("value"       )  @Expose var value       : Int?    = null,
-  @SerializedName("quantity"    )  @Expose var quantity    : Int?    = null,
-  @SerializedName("weight"      )  @Expose var weight      : Int?    = null
+  @SerializedName("value"       )  @Expose var value       : Double?    = null,
+  @SerializedName("quantity"    )  @Expose var quantity    : Double?    = null,
+  @SerializedName("weight"      )  @Expose var weight      : Double?    = null
 
 )
