+++ terminal_androidsdk/src/main/java/com/terminal/terminal_androidsdk/core/model/component_shipment/Parcel.kt
@@ -6,16 +6,16 @@ import com.google.gson.annotations.SerializedName
 
 data class Parcel (
 
-  @SerializedName("_id"          )  @Expose var Id          : String?          = null,
-  @SerializedName("description"  )  @Expose var description : String?          = null,
-  @SerializedName("items"        )  @Expose var items       : ArrayList<Items> = arrayListOf(),
-  @SerializedName("packaging"    )  @Expose var packaging   : Packaging?       = Packaging(),
-  @SerializedName("total_weight" )  @Expose var totalWeight : Int?             = null,
-  @SerializedName("user"         )  @Expose var user        : String?          = null,
-  @SerializedName("weight_unit"  )  @Expose var weightUnit  : String?          = null,
-  @SerializedName("parcel_id"    )  @Expose var parcelId    : String?          = null,
-  @SerializedName("created_at"   )  @Expose var createdAt   : String?          = null,
-  @SerializedName("updated_at"   )  @Expose var updatedAt   : String?          = null,
-  @SerializedName("__v"          )  @Expose var _v          : Int?             = null
+  @SerializedName("_id")  @Expose var Id          : String?          = null,
+  @SerializedName("description")  @Expose var description : String?          = null,
+  @SerializedName("items")  @Expose var items       : ArrayList<Items> = arrayListOf(),
+  @SerializedName("packaging")  @Expose var packaging   : Packaging?       = Packaging(),
+  @SerializedName("total_weight")  @Expose var totalWeight : Double?             = null,
+  @SerializedName("user")  @Expose var user        : String?          = null,
+  @SerializedName("weight_unit")  @Expose var weightUnit  : String?          = null,
+  @SerializedName("parcel_id")  @Expose var parcelId    : String?          = null,
+  @SerializedName("created_at")  @Expose var createdAt   : String?          = null,
+  @SerializedName("updated_at")  @Expose var updatedAt   : String?          = null,
+  @SerializedName("__v")  @Expose var _v          : Double?             = null
 
 )
