+++ terminal_androidsdk/src/main/java/com/terminal/terminal_androidsdk/core/model/component_shipment/Quantity.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Quantity (
 
-  @SerializedName("value"             )  @Expose var value             : Int?    = null,
+  @SerializedName("value"             )  @Expose var value             : Double?    = null,
   @SerializedName("unitOfMeasurement" )  @Expose var unitOfMeasurement : String? = null
 
 )
